requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - COMPATIBLE VERSIONS
asyncpg==0.29.0
databases[postgresql]==0.9.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Image Processing & PDF Generation
Pillow==10.2.0
img2pdf==0.5.1

# HTTP Client
httpx==0.26.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Templates (Jinja2)
jinja2==3.1.3